lib-storage: mailbox_get_path() and friends can now return an error. An upcoming "index" mailbox_list backend uses mailbox GUIDs as the directory names. Doing a name => GUID lookup may fail, so the path lookup may fail.
This is a somewhat annoying change and perhaps it can be removed in future,
but for now it looks like it can't be avoided.

+++ src/lib-storage/list/mailbox-list-delete.c
@@ -19,7 +19,7 @@ mailbox_list_check_root_delete(struct mailbox_list *list, const char *name,
 {
 	const char *root_dir;
 
-	root_dir = mailbox_list_get_root_path(list, MAILBOX_LIST_PATH_TYPE_DIR);
+	root_dir = mailbox_list_get_root_forced(list, MAILBOX_LIST_PATH_TYPE_DIR);
 	if (strcmp(root_dir, path) != 0)
 		return 0;
 
@@ -51,7 +51,9 @@ int mailbox_list_delete_maildir_via_trash(struct mailbox_list *list,
 	const char *src, *trash_dest;
 	unsigned int count;
 
-	src = mailbox_list_get_path(list, name, MAILBOX_LIST_PATH_TYPE_MAILBOX);
+	if (mailbox_list_get_path(list, name, MAILBOX_LIST_PATH_TYPE_MAILBOX,
+				  &src) <= 0)
+		i_unreached();
 	if (mailbox_list_check_root_delete(list, name, src) < 0)
 		return -1;
 
@@ -120,9 +122,13 @@ int mailbox_list_delete_mailbox_file(struct mailbox_list *list,
 				     const char *name)
 {
 	const char *path;
+	int ret;
 
-	path = mailbox_list_get_path(list, name,
-				     MAILBOX_LIST_PATH_TYPE_MAILBOX);
+	ret = mailbox_list_get_path(list, name, MAILBOX_LIST_PATH_TYPE_MAILBOX,
+				    &path);
+	if (ret < 0)
+		return -1;
+	i_assert(ret > 0);
 
 	/* we can simply unlink() the file */
 	if (unlink(path) == 0)
@@ -241,7 +247,7 @@ void mailbox_list_delete_until_root(struct mailbox_list *list, const char *path,
 	const char *root_dir, *p;
 	unsigned int len;
 
-	root_dir = mailbox_list_get_root_path(list, type);
+	root_dir = mailbox_list_get_root_forced(list, type);
 	if (strncmp(path, root_dir, strlen(root_dir)) != 0) {
 		/* mbox workaround: name=child/box, root_dir=mail/.imap/,
 		   path=mail/child/.imap/box. we'll want to try to delete
@@ -281,10 +287,10 @@ static void mailbox_list_try_delete(struct mailbox_list *list, const char *name,
 {
 	const char *mailbox_path, *path;
 
-	mailbox_path = mailbox_list_get_path(list, name,
-					     MAILBOX_LIST_PATH_TYPE_MAILBOX);
-	path = mailbox_list_get_path(list, name, type);
-	if (path == NULL || *path == '\0' || strcmp(path, mailbox_path) == 0)
+	if (mailbox_list_get_path(list, name, MAILBOX_LIST_PATH_TYPE_MAILBOX,
+				  &mailbox_path) <= 0 ||
+	    mailbox_list_get_path(list, name, type, &path) <= 0 ||
+	    strcmp(path, mailbox_path) == 0)
 		return;
 
 	if (*list->set.maildir_name == '\0' &&
@@ -332,8 +338,14 @@ int mailbox_list_delete_symlink_default(struct mailbox_list *list,
 					const char *name)
 {
 	const char *path;
+	int ret;
 
-	path = mailbox_list_get_path(list, name, MAILBOX_LIST_PATH_TYPE_DIR);
+	ret = mailbox_list_get_path(list, name, MAILBOX_LIST_PATH_TYPE_DIR,
+				    &path);
+	if (ret < 0)
+		return -1;
+	i_assert(ret > 0);
+
 	if (unlink(path) == 0)
 		return 0;
